requirements.txt
feedparser
Newspaper3k
requests
google-api-python-client
progressbar2
ffmpy